fta_backend/requirements.txt
Django==5.1.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
djangorestframework==3.15.2
python-dotenv==1.0.1
django-cors-headers==4.7.0